\name{Abalone}
\docType{data}
\alias{Abalone}
\alias{AbaDF}
\alias{AbUnits}
\alias{AbaloneIdt}
\title{Abalone Data Set}
\description{A interval-valued data set containing 24 units, created from from the Abalone dataset (UCI Machine Learning Repository), after aggregating by sex and age.}
\usage{data(Abalone)}
\format{AbdaDF: A data frame containing the original 4177 Abalone individuals described by 7 variables.\cr
       AbUnits: A factor with 4177 observations and 24 levels indicating the sex by age combination to which each orginal individual belongs to.\cr
       AbaloneIdt: An IData object with 24 observations and 7 interval-valued variables, describing the intervals formed by aggregating the AbdaDF microdata by the AbUnits factor.}
\keyword{datasets}
\keyword{interval data}


