\name{Cars}
\docType{data}
\alias{Cars}
\title{Cars Data Set}

\description{This data set consist of the intervals for four characteristics (Price, EngineCapacity, TopSpeed and Acceleration) of 27 cars models partitioned into four different classes (Utilitarian, Berlina, Sportive and Luxury).}

\usage{data(Cars)}

\format{A data frame containing 27 observations on 9 variables, the first eight with the the lower and upper bounds of the interval characteristics for 27 car models, the last one a factor indicating the model class.}

\keyword{datasets}
\keyword{interval data}


