\name{CarsData}
\docType{data}
\alias{CarsData}
\title{Cars Data Set}

\description{This data set consist of MidPoints and LogRanges of the intervals for four characteristics (Price, EngineCapacity, TopSpeed and Acceleration) of 27 cars models partitioned into four diffetent classes (Utilitarian,Berlina,Sportive and Luxury).}

\usage{data(CarsData)}

\format{A data frame containing 27 observations on 9 variables, the first eight with the MidPoint and LogRange of the interval characteristics for 27 car models, the last one a factor indicating the model class.}

\keyword{datasets}
\keyword{interval data}


