\name{ChinaTemp}
\docType{data}
\alias{ChinaTemp}
\title{China Temperatures Data Set}
\description{This data set consist of the intervals of observed temperatures (Celsius scale) in each of the four quarters, Q_1 to Q_4, of the years 1974 to 1988 in 60 chinese meteorologic stations; one outlier observation (YinChuan_1982) has been discarded. 
The 60 stations belong to different regions in China, which therefore define a partition of the 899 stations-year combinations.}
\usage{data(ChinaTemp)}
\format{A data frame containing 899 observations on 9 variables, the first eight with the lower and upper bounds of the temperatures by quarter in the 899 stations-year combinations, the last one a factor indicating the geographic region of each station.}
\keyword{datasets}
\keyword{interval data}


