\name{extmatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{extmatrix-class}

\title{Class \dQuote{extmatrix}}

\description{\dQuote{extmatrix} is a simple extension of the base matrix class, that that accepts NULL objects as members.}

\section{Extends}{ Class \code{\link{matrix}}, directly.}




