\name{IdtMxtNDE-class}
\Rdversion{1.1}
\docType{class}
\alias{IdtMxtNDE-class}

\title{Class IdtMxtNDE}

\description{IdtMxtNDE is an union of classes \code{\linkS4class{IdtMxNDE}} and \code{\linkS4class{IdtMxNDRE}}, containing the results of mixture Normal model parameter estimation by maximum likelihood (\code{\linkS4class{IdtMxNDE}}) or robust (\code{\linkS4class{IdtMxNDRE}}) methods.}

\seealso{
   \code{\linkS4class{IdtE}}, \code{\linkS4class{IdtMxE}},  \code{\linkS4class{IdtMxNDE}}, \code{\linkS4class{IdtMxNDRE}} 
}

\keyword{classes}
\keyword{interval data}
