\name{MAINT.Data-internal}

\alias{print.summaryIData}
\alias{==.IData}
\alias{!=.IData}
\alias{[.IData}
\alias{[<-.IData}

\alias{show,IData-method}
\alias{nrow,IData-method}
\alias{ncol,IData-method}
\alias{coef,IdtNDE-method}
\alias{stdEr,IdtNDE-method}
\alias{RepLOptim}

\title{Operators and Internal MAINT.Data functions.}
\description{Internal MAINT.Data functions and standard operators.}
\keyword{internal}
