\name{MAINT.Data-internal}

\alias{SigEf_ps}
\alias{SigEf_alphas}
\alias{SigEf_betas}
\alias{SigEf_gammas}

\alias{==,IData,IData-method}
\alias{!=,IData,IData-method}
\alias{[,IData-method}
\alias{[<-,IData-method}
\alias{rbind,IData-method}
\alias{cbind,IData-method}

\alias{show,IData-method}
\alias{nrow,IData-method}
\alias{ncol,IData-method}
\alias{dim,IData-method}
\alias{rownames,IData-method}
\alias{colnames,IData-method}
\alias{names,IData-method}

\alias{summary,IdtMANOVA-method}
\alias{show,IdtOutl-method}

\alias{MidPoints}
\alias{LogRanges}
\alias{Ranges}
\alias{NbMicroUnits}

\alias{CovCase-methods}
\alias{CovCase}

\alias{MCDcnp2}
\alias{MCDRewcnp2}
\alias{C1CPgrad}
\alias{C1CovPar}
\alias{C2CPgrad}
\alias{C2CovPar}
\alias{C2GetCov}
\alias{C2GetCovStderr}
\alias{C2RestCovInd}
\alias{C3CPgrad}
\alias{C3CovPar}
\alias{C3RestCovInd}
\alias{C4CPgrad}
\alias{C4CovPar}
\alias{C4RestCovInd}
\alias{C5CPgrad}
\alias{C5RestCovInd}
\alias{Cnf2MaxLik}
\alias{CovC1toRestCov}
\alias{CovtoParC2}
\alias{EvalClrule}
\alias{FreePar}
\alias{GC2LogLikC.grad}
\alias{GC2mLogLik}
\alias{GC2mLogLik.grad}
\alias{GetCovPar}
\alias{Getvarofsmplmeansbygrp}
\alias{ILogLikDNC}
\alias{ILogLikNC}
\alias{ILogLikNC1}
\alias{IdtFDMxtSNmle}
\alias{IdtHetMxtNmle}
\alias{IdtNmle}
\alias{IdtSNmle}
\alias{MINPHIARG}
\alias{MxtDist.fit}
\alias{NestedBy}
\alias{NextModel}
\alias{ConfNestedBy}
\alias{ConfNextModel}
\alias{getConfig}
\alias{ObsLik.Omegagrad}
\alias{ObsLik.alphagrad}
\alias{ObsLik.ksigrad}
\alias{ObsLogLiks}
\alias{NmleVCov}
\alias{ObsLogLiks,IdtSngNDE-method}
\alias{OmgInv.grad}
\alias{OmgtoOmgbar.grad}
\alias{PartoMatC2}
\alias{RestCov}
\alias{RestCov.grad}
\alias{RestCovInd}
\alias{RestCovj}
\alias{Rfasttle}
\alias{Rfulltle}
\alias{RunfooforallComb}
\alias{SKC1GetStderr}
\alias{SKGetStderr}
\alias{SKnpar}
\alias{SNCMaxLik}
\alias{SNCnf1MaxLik}
\alias{Sigmagrad}
\alias{SigmagradCnt}
\alias{SigmasrgradCnt}
\alias{Subsmplmle}
\alias{Theta.ll.grad}
\alias{VCovGC2LogLikC.grad}
\alias{b}
\alias{b0}
\alias{b2}
\alias{c2cind}
\alias{c2rfind}
\alias{cfsbmu0}
\alias{cnvCPtoDP}
\alias{cnvDPtoCP}
\alias{dvec0ind}
\alias{dvecind}
\alias{error}
\alias{fasttle1}
\alias{fulltle1}
\alias{GetMD2}
\alias{fL}
\alias{fL.grad}
\alias{fL2}
\alias{fmdind}
\alias{getfskpar}
\alias{initparconf2}
\alias{lftD}
\alias{lftDplus}
\alias{ln2pi}
\alias{log2Phi}
\alias{ltdind}
\alias{ltind0}
\alias{ltind1}
\alias{lttoutind}
\alias{MDOtlDet}
\alias{mleNSigC35}
\alias{mleVCov}
\alias{msn.dev.grad}
\alias{msnCP.dev}
\alias{msnCP.dev.grad}
\alias{msnCP.ll.grad}
\alias{msnDP.ll.grad}
\alias{mymsn.dev}
\alias{mymsn.dev.grad}
\alias{mymsn.mle}
\alias{mymsn.moment.fit}
\alias{ncovp}
\alias{npar}
\alias{pdwt.solve}
\alias{phioverPhi}
\alias{refinementstep}
\alias{rghtD}
\alias{safepdwt.solve}
\alias{Ilocsnda}
\alias{Igensnda}
\alias{trialstep}
\alias{ulbeta02}
\alias{utdind}
\alias{utind0}
\alias{utind1}
\alias{uttoltind}
\alias{zeta1}
\alias{Ilda}
\alias{Iqda}

\alias{CPtoDP.ll.grad}
\alias{DPll.grad}
\alias{HetSNmleVCov}
\alias{SKC1Getvcov}
\alias{SKGetvcov}
\alias{SNVCovscaling}
\alias{SigCind}
\alias{mleNvcovC35}
\alias{vcovCind}
\alias{resdlmfit}

\alias{delta.etc} 
\alias{mydp2cpMv} 
\alias{force.symmetry} 
\alias{msn.dp2cp} 
\alias{msn.moment.fit}
\alias{mysn.infoMv}

\alias{casy}
\alias{CHmasy}
\alias{HDmpred}

\alias{summaryIData-class}
\alias{summaryIdtMclust-class}
\alias{show,summaryIData-method}
\alias{show,summaryIdtMclust-method}

\title{Operators, functions and other Internal MAINT.Data objects.}
\description{Internal MAINT.Data objects.}
\keyword{internal}
